feat(renderer): render whole document
Modify the grammar to return the document as a tree
of blocks (list, image blocks, paragraphs, delimited blocks),
which also include sections, which can themselves contain
further blocks. This is a change to the previous approach in
which the document was seen as a list of (erroneously named)
'lines'.

Also:
- fix line-return on blank lines (not ignored anymore)
and lists
- move code to `section_test.go` since headings are not
block-level elements, and fix heading templates based on
level (1, 2 or more)

Signed-off-by: Xavier Coulon <xcoulon@redhat.com>