fix(parser): fix admonition paragraph parsing
Parsing failed if the `[<kind>]` attribute was not followed by a space
the reason is that is was considered as a regular attribute instead of
an admonition marker, which resulted in a regular paragraph instead of
an admonition paragraph. The side-cause if that regular attributes did
not expect spaces between `]` and the NEWLINE, which needs to be supported,
too.

Fixes #88

Signed-off-by: Xavier Coulon <xcoulon@redhat.com>