fix(parser): avoid too much parsing for sections
fix some rules to avoid too much parsing. For example, just
look for the next section title at the end of a section instead
of a parsing the whole next section(s), which is exponentially
expensive.

also, use an optmized grammar, which means that some entrypoint
cannot be used during the tests (they are refactored or moved into
the single parent rule that uses them).

also, restore blanklines in resulting AST, but keep ignoring
during the rendering phase

Fixes @121

Signed-off-by: Xavier Coulon <xcoulon@redhat.com>