test(fixtures): add fixtures testing framework (#221)
Create a basic framework for adding test fixtures. The workflow is to
add .adoc files to the fixtures folder with corresponding .html files to
act as the "golden" HTML output. Include files are in the includes
sub-folder so as to differentiate them from the main fixture files.

Failed tests show a diff against the golden output.

Tests are currently hidden behind a "fixtures" build tag. Use
"-tags=fixtures" when running tests to activate.

The existing test fixtures are mostly taken from the asciidoctor test
fixtures. The golden files were generated with asciidoctor v1.5.7.1
using the "-s" option.

The `test-fixtures` goal will run ginkgo with the "fixtures" build tag and focus only
on specs with descriptions contains "fixtures."

Fixes #79