feat(parser/renderer): support file inclusions
support file inclusions in sections and at the
root of the document.
Introduced a "pre-rendering" phase in which file
inclusions are processed, then preamble and table
of contents macro are inserted (if applicable), which
means that the result of the parsing does not include
the preamble nor the ToC macro anymore.

fixes #307

Signed-off-by: Xavier Coulon <xcoulon@redhat.com>