refactor(parser/renderer): refactor parsing of document header (#338)
Include the section with level 0 as the top root element of the document (other section with level 0 can follow, though, but in the future, such a thing will be treated as an error)

Also, refactor document authors and revision when set after document header (section 0 when it is the first element). Authors and revisions are now parsed in a more simple manner and set as attributes of the header/section. The authors are more finely "processed" during rendering.
also, exclude all content from the document attributes during the parsing phase. As
a consequence, document attribute declarations are part of the resulting DOM as regular
elements.

Fixes #335

Signed-off-by: Xavier Coulon xcoulon@redhat.com