refactor(parser/renderer): include files before full parsing
add a new entrypoint in the grammar to "preparse" the document, only
looking for a subset of elements, in particular the `include` directives
and the sections whose level may be offset during the file inclusions.
Once the file have been included (ie, the `include` directives have been
replaced with the actual file content), then the document can be fully parsed.

Fixes #343

Signed-off-by: Xavier Coulon <xcoulon@redhat.com>