refactor(types/parser/renderer): refactor grammar
- simpler grammar that does not resolve the nesting of
  sections, nor the items in lists.
- the result of the parsing with this new grammar is that's called
  a "preflight document" (until a better name is found), which contains
  all the blocks, but not "rearranged" in sections, lists, etc. It
  also has some "file inclusion" elements to be resolved.
- the final "document" is processed from this "preflight doc", where
  all sections, lists, file inclusions, table of contents, preamble
  are put in their expected place.

Also, use predicate in rules to limit the number of occurrences of a
character ("=" in section titles and "." in ordered list items)

Also, this approach avoids the need for "preparsing" the document and resolving
the file inclusions, as it was done before.

Also: using pointers to blocks to avoid passing the values to functions,
which might become expensive when a document has multiple large paragraphs.

Also, many more tests since in the 'pkg/parser' pkg, where the "preflight doc"
and then the final "document" are evaluated.

Fixes #362

Signed-off-by: Xavier Coulon <xcoulon@redhat.com>