refactor(parser): apply attribute substitutions in final Document
DocumentAttributeDeclaration, Substitution and Reset
blocks are processed while preparing the final
document.
As a consequence, final document is simpler to render,
but more processing needs to take place beforehand,
including the section id generation, the implicit usage of
the  'imagesdir' attribute while generating the image
location and the marker for the table of contents.

If an attribute was not declared and could not be sustituted, then
it is replace with its name surrounded by curly brackets.

fixes #444

Signed-off-by: Xavier Coulon <xcoulon@redhat.com>