feat(renderer): return list of sections after rendering
The hierarchical list of sections is returned from the call
to the `libasciidoc.ConvertToHTML(...)` and
`libasciidoc.ConvertFileToHTML(...)` funcs.
The `Title` field is the HTML displayed in the table of
contents (if it were displayed).
The levels of the returned table of contents is `2` by
default, unless a `:toclevels:` attribute was specified
in the document.

Fixes #363

Signed-off-by: Xavier Coulon <xcoulon@redhat.com>