refactor(parser): refactor delimited blocks
parse the content of the delimited blocks
to detect file inclusions and apply accordingly.
The result is a set of `VerbatimLine` elements
which are then converted back into a document
which is parsed with the grammar that corresponds
to the type of delimited block.

also: ignore errors about dot imports on Gomega/Ginkgo

Fixes #62
Fixes #455

Signed-off-by: Xavier Coulon <xcoulon@redhat.com>