Refactor bench and profile make targets (#589)
Create a separate `profile` target for profiling. Only run the
benchmarks once in this target.

Update `bench` target to make five passes and save the output to a file
for use with `benchstat`.

Add a `bench-smoke` target to smoke test the top-level benchmarks by
running them only once.  This target is run via CI to ensure that the
benchmarks are functional.

Fixes #580