feat(parser/renderer/types): Support inline and image icons
This adds inline icon support, and significantly refactors the icon handling
to bring in support for image based icons and all of the icon attributes
supported by asciidoctor, bringing us to feature parity.

Additionally, we preserve some of the attributes from the font icon tags onto
images, allowing flip, rotate, and size styles to be applied to icons.
(This is an improvement over asciidoctor, and requires supporting CSS.)

One subtle difference when comparing output.  We emit source paths for images
without a leading "./".  That is because our use of path.Join() automatically
removes the redundant portion.

Fixes #587
Fixes #611
Fixes #610