refactor(renderer): Use inline UTF-8 (optionally) (#645)
This adds an option to render UTF-8 as is, doing the final
conversion to HTML entities only at the final step.  This
is more likely to result in correct encoding as they remain
unadulterated until the final rendering step.

Furthermore, some renderers may wish to allow Unicode to
be encoded as is.  This can be set on the document context,
but we also set it by default when a new "unicode" document
attribute is present.

Fixes #626