feature(parser/renderer): Render curved quotes (#648)
This provides support for both explicit typographical quotes, and
the implicit curved apostrophe when it appears between letters
as in mother's day.

This also refactors the handling of certain typographic symbols,
moving their handing from the render to the parser, so that they
are treated properly during the parse phase.  (This was important
in properly handling these curved quotes.)

Also, the greedy rules for string elements in quoted text were
removed, as they were interfering with the parse, and not strictly
required.  (They were an optimization.)  It's our opinion that for
most documents quoted text will be in the minority, such that
the performance gain provided by the old rules was minimal at best.

Fixes #176