refactor(parser): Refactor ordered lists for multiple roles
This adds support for the multiple role markup to unordered
lists.  This also cleans up the NumberingStyle, using just the
normal Style attribute.

Finally a subtle bug was fixed in the parser for dealing
with PositionalValues, which we had to fix here in order to
pass the various tests for ordered lists.

There are also a few more tests added to cover that case, in
other areas, written while diagnosing that bug.