bug(parser): Inline attribute parsing is incorrect
(Note that block attributes, and some other inline attributes,
suffer from this problem -- this PR only fixes icon and image
inline attributes.)

While working on some new support for align and float attributes,
we noticed that the parser was fragile, and failed on certain
cases including double embedded whitespace, embedded quote marks,
and others.

This logic will be expanded to other elements and to block
attributes; this incremental PR should be easier to review.