features(types/renderer): Table cols attribute support (#698)
This adds support for handling the cols attribute ont tables,
giving the ability specify per column widths and alignments.
It also supports the repeat syntax, and the ability to set
autoexpand (either for the entire table or for an individual row).

The width handling and calculations were moved to instantiation
time for the Table, so that all backends can benefit from the
calculation work done for these.

The style component of the column specification is consumed and
stored, but does not affect deeper parsing or rendering yet.

This takes table support about as far as it can go without
reworking the parser to provide context-sensitive parsing of
the table contents.

While here the limitations document was updated to provide more
detail on limitations, including links to issues for folks
who want to track progress.

Fixes #694
Fixes #686