refactor(renderer) Newlines in should be in templates, not renderer
This refactors the renderer somewhat so that newlines are no
longer injected in between block elements by the renderer,
but rather the closing block element has a newline appended
in the closing block tag in the template.

This moves the control over these newlines to the templates,
and makes the logic to understand when they get added easier
to understand.

This affects pretty much all the tests that look for HTML output,
as they now have a newline at the end (unless they are enclosed
within full html tags.)  For the most part, they just had a newline
appended their expected output.

Note that this does not introduce stray whitespace in the *middle*
of documents.

Two places where stray newlines were introduced (possibly for
compatibility with asciidoctor) was chagned -- if the document body
is empty then we do not emit a blank line.  Likewise we do not do
so if a blockquote has no body content.

This will not have any effect on rendered output.

Fixes #699