feature(renderer): version-label not honored
This supports the version-label be changed, or even reset
for document attributes.

To support resetting this label, the initial value is located
in the Predefined attributes, and a reset of the attribute is
stored as a nil attribute value instead of deleting it from the
attributes.  This allows us to distinguish a reset as an override.

This approach will be used for other predefined text values in
follow ups.

Fixes #710