feature(renderer): Tables should honor table-caption attribute
This uses a kind of hack to make Table captions work mostly as
they are documented to, so that {counter:table-number} seems to
act like a substitution.  This hack should be removed when full
attribute value substitution is done.  (This approach has the
benefit of localizing this hack until it can be replaced.)

Of course, this will be incompatible with other uses of the
table-counter counter (because it isn't actually the same counter),
and won't support anything beyond the most trivial substitutions.

While here removed a couple of unused functions.