bug(renderer): source highlighter should get unadulterated string
Ths fix for this involved some refactoring of the highlighting work
to separate out the calloouts (we embed them using reserved Unicode so
that we can find them again and replace them inline).

Chroma automatically escapes the code for HTML safetly already, so
we defer to it's rules, when we use it.  We still have to do this
ourselves when not using Chroma.

Fixes #721