feature(renderer): Customizable admonition captions
This allows document attributes to replace the default captions
used for admonitions, as well as honoring the "caption" icon when
used locally on an admonition.  It also corrects some errors in the
handling of inline icons, where alt tags should not be capitalized,
and for admonition icons which should never carry an extraneous title.

Also, admonition icons were not getting the attributes sent down
to permit them to be adjusted properly, so that is fixed as well.

Results from the various tests were compared with asciidoctor.

Fixes #679