refactor(parser/renderer): special characters at parser level
Parsing special characters ('>', '<', '&') and also
'+' (for compatibility of output with Asciidoctor)

These special characters are stored in the new `SpecialCharacter`
type and can be converted into HTML entities in the HTML5
and XHTML5 backends. Other backends may decide to display
the values as-is.
This removes the introduction of HTML entities during
attribute substitutions (which would cause troubles for
future backends) and also, removes the need to
the `sanitized` type (an alias for `html.HTML`)

Fixes #729

Signed-off-by: Xavier Coulon <xcoulon@redhat.com>