refactor(parser/renderer): special characters at parser level (#734)
Parsing special characters (`>`, `<`, `&`) as well as
`+` for output compatibility with Asciidoctor

These special characters are stored in the new `SpecialCharacter`
type and can be converted into HTML entities by the HTML5
and XHTML5 backends. 
Other backends may decide to display the values as-is.

This removes the usage of HTML entities during
attribute substitutions at the parser level, which could 
cause troubles for future, non-HTML backends. 

Also, it removes the need to the `sanitized` type 
(an alias for `html.HTML`)

Fixes #729

Signed-off-by: Xavier Coulon <xcoulon@redhat.com>