chore(build): Use go generate
This provides a simpler go generate (meaning developers no longer
need to use "make", which is a qualify of life improvement for
developers working on non-UNIX systems), and updates the instructions
to reflect the trivial installation steps. (Also details about using
go generate for developers are added to the CONTRIBUTING document.)

This also coincidentally includes updates to the latest dependencies,
which has the benefit of eliminating the protobuf dependency.

Fixes #638