feature(parser): custom subs on paragraphs
Support for one or more custom substitutions
Starting with Rawline elements which are parsed
and result in other elements. Subsequent substitutions
only apply on the StringElements.
At the end of the substitutions, the lines are preserved

Also, fix issues with single line comments which should not
start with spaces before the `//` marker

Fixes #597

Signed-off-by: Xavier Coulon <xcoulon@redhat.com>