feat(parser): support custom substitutions on delimited blocks
applying the same logic as with paragraphs to apply default and
custom substitutions on delimited blocks. This means that
delimited blocks are processed in multiple steps, but each step
has a much simpler grammar.

Also, quite a lot of refactoring and code reorganization for the
delimited block tests in `pkg/parser`.

See https://github.com/bytesparadise/libasciidoc/pull/749

Fixes #558

Signed-off-by: Xavier Coulon <xcoulon@redhat.com>