refactor(parser): parse attributes less often
when processing blocks at the document level or
or within certain kind of delimited blocks, parse
the attributes once and store them in a dedicated stack
in the global store (even if `nil`).
Then  retrieve from the stack in the global store
when initializing the block and finally, remove from
the stack.

Signed-off-by: Xavier Coulon <xcoulon@redhat.com>